Introduce sync_on_lookup option off by default
This option, which is off by default, when enabled, davfs2 will
interrogate the server about file sizes to catch cases when a file on
the server has been updated.

It turns out that for most users this introduces performance issue, so
this option has been introduced to solve the issue.

Fixes #20